roxygen tags for data and compact data description table
The current output of makeOxygen, when applied to a dataframe, adds roxygen tags for @export, @rdname and @examples. These are either not relevant of not advised for documenting data. (at least according to Hadley http://r-pkgs.had.co.nz/data.html#documenting-data). I suggest changing "add_fields" to  add_fields <- c("details", "source"). I may not have suggested the best way to do this as I did not use sinew_opts, and basically overwrite add_fields if the object type is data.frame or tibble.

Also use a tabular format for the data description rather than \describe{} as it is much more compact.